Bug: Inspections warnings "module xyz not found"
In the inspections were several warnings, that the module that should be parsed couldn't be found.
Reason was that sometimes Pylint received relative paths as input, which of course couldn't be parsed.
Underlying cause of this was that the given virtual file was a deprecated VirtualFileWindowsImpl.
Hacky workaround now retrieves the original virtual file with the getDelegate() method.